DLA: fix cmpNodes in DeduplicateFields
Before this commit, cmpNodes was only using the number of successors to
decide if two nodes could be equivalent or to bail out.
This was wrong because it did not take into consideration the kind of
successors.
Hence it failed to distinguish e.g. the case where both nodes had a
single successor, but one was a pointer successor and the other an
instance successor.
This commit fixes the problem, properly considering the kind of
successors.